Refactor: Creates epics folder for epics and move back actioncreators
