+++ public/app/features/explore/state/actionTypes.ts
@@ -1,14 +1,25 @@
 // Types
 import { Emitter } from 'app/core/core';
 import {
+  RawTimeRange,
   DataQuery,
   DataSourceSelectItem,
   DataSourceApi,
   QueryFixAction,
   LogLevel,
   TimeRange,
+  DataQueryError,
+  SeriesData,
 } from '@grafana/ui/src/types';
-import { ExploreId, ExploreItemState, HistoryItem, RangeScanner, ExploreUIState, ExploreMode } from 'app/types/explore';
+import {
+  ExploreId,
+  ExploreItemState,
+  HistoryItem,
+  RangeScanner,
+  ExploreUIState,
+  ExploreMode,
+  QueryOptions,
+} from 'app/types/explore';
 import { actionCreatorFactory, noPayloadActionCreatorFactory, ActionOf } from 'app/core/redux/actionCreatorFactory';
 
 /**  Higher order actions
@@ -128,6 +139,21 @@ export interface ModifyQueriesPayload {
   modifier: (query: DataQuery, modification: QueryFixAction) => DataQuery;
 }
 
+export interface QueryFailurePayload {
+  exploreId: ExploreId;
+  response: DataQueryError;
+}
+
+export interface QueryStartPayload {
+  exploreId: ExploreId;
+}
+
+export interface QuerySuccessPayload {
+  exploreId: ExploreId;
+  result: any;
+  latency: number;
+}
+
 export interface HistoryUpdatedPayload {
   exploreId: ExploreId;
   history: HistoryItem[];
@@ -141,6 +167,15 @@ export interface RemoveQueryRowPayload {
 export interface ScanStartPayload {
   exploreId: ExploreId;
   scanner: RangeScanner;
+}
+
+export interface ScanRangePayload {
+  exploreId: ExploreId;
+  range: RawTimeRange;
+}
+
+export interface ScanStopPayload {
+  exploreId: ExploreId;
 }
 
 export interface SetQueriesPayload {
@@ -188,7 +223,51 @@ export interface LoadExploreDataSourcesPayload {
   exploreDatasources: DataSourceSelectItem[];
 }
 
+export interface RunQueriesPayload {
+  exploreId: ExploreId;
+}
+
+export interface ResetQueryErrorPayload {
+  exploreId: ExploreId;
+  refIds: string[];
+}
+
 export interface SetUrlReplacedPayload {
+  exploreId: ExploreId;
+}
+
+export interface ProcessQueryErrorsPayload {
+  exploreId: ExploreId;
+  response: any;
+  datasourceId: string;
+}
+
+export interface ProcessQueryResultsPayload {
+  exploreId: ExploreId;
+  response: any;
+  latency: number;
+  datasourceId: string;
+}
+
+export interface ProcessQueryResultsPayload {
+  exploreId: ExploreId;
+  response: any;
+  latency: number;
+  datasourceId: string;
+}
+
+export interface RunQueriesBatchPayload {
+  exploreId: ExploreId;
+  queryOptions: QueryOptions;
+}
+
+export interface SubscriptionDataReceivedPayload {
+  data: SeriesData;
+  exploreId: ExploreId;
+}
+
+export interface LimitMessageRatePayload {
+  data: SeriesData;
   exploreId: ExploreId;
 }
 
@@ -282,9 +361,32 @@ export const loadDatasourceReadyAction = actionCreatorFactory<LoadDatasourceRead
 export const modifyQueriesAction = actionCreatorFactory<ModifyQueriesPayload>('explore/MODIFY_QUERIES').create();
 
 /**
+ * Mark a query transaction as failed with an error extracted from the query response.
+ * The transaction will be marked as `done`.
+ */
+export const queryFailureAction = actionCreatorFactory<QueryFailurePayload>('explore/QUERY_FAILURE').create();
+
+export const queryStartAction = actionCreatorFactory<QueryStartPayload>('explore/QUERY_START').create();
+
+/**
+ * Complete a query transaction, mark the transaction as `done` and store query state in URL.
+ * If the transaction was started by a scanner, it keeps on scanning for more results.
+ * Side-effect: the query is stored in localStorage.
+ * @param exploreId Explore area
+ * @param transactionId ID
+ * @param result Response from `datasourceInstance.query()`
+ * @param latency Duration between request and response
+ * @param queries Queries from all query rows
+ * @param datasourceId Origin datasource instance, used to discard results if current datasource is different
+ */
+export const querySuccessAction = actionCreatorFactory<QuerySuccessPayload>('explore/QUERY_SUCCESS').create();
+
+/**
  * Remove query row of the given index, as well as associated query results.
  */
 export const removeQueryRowAction = actionCreatorFactory<RemoveQueryRowPayload>('explore/REMOVE_QUERY_ROW').create();
+
+export const runQueriesAction = actionCreatorFactory<RunQueriesPayload>('explore/RUN_QUERIES').create();
 
 /**
  * Start a scan for more results using the given scanner.
@@ -292,6 +394,12 @@ export const removeQueryRowAction = actionCreatorFactory<RemoveQueryRowPayload>(
  * @param scanner Function that a) returns a new time range and b) triggers a query run for the new range
  */
 export const scanStartAction = actionCreatorFactory<ScanStartPayload>('explore/SCAN_START').create();
+export const scanRangeAction = actionCreatorFactory<ScanRangePayload>('explore/SCAN_RANGE').create();
+
+/**
+ * Stop any scanning for more results.
+ */
+export const scanStopAction = actionCreatorFactory<ScanStopPayload>('explore/SCAN_STOP').create();
 
 /**
  * Reset queries to the given queries. Any modifications will be discarded.
@@ -310,6 +418,8 @@ export const splitCloseAction = actionCreatorFactory<SplitCloseActionPayload>('e
  * The copy keeps all query modifications but wipes the query results.
  */
 export const splitOpenAction = actionCreatorFactory<SplitOpenPayload>('explore/SPLIT_OPEN').create();
+
+export const stateSaveAction = noPayloadActionCreatorFactory('explore/STATE_SAVE').create();
 
 /**
  * Update state of Explores UI elements (panels visiblity and deduplication  strategy)
@@ -355,7 +465,27 @@ export const loadExploreDatasources = actionCreatorFactory<LoadExploreDataSource
 
 export const historyUpdatedAction = actionCreatorFactory<HistoryUpdatedPayload>('explore/HISTORY_UPDATED').create();
 
+export const resetQueryErrorAction = actionCreatorFactory<ResetQueryErrorPayload>('explore/RESET_QUERY_ERROR').create();
+
 export const setUrlReplacedAction = actionCreatorFactory<SetUrlReplacedPayload>('explore/SET_URL_REPLACED').create();
+
+export const processQueryErrorsAction = actionCreatorFactory<ProcessQueryErrorsPayload>(
+  'explore/PROCESS_QUERY_ERRORS'
+).create();
+
+export const processQueryResultsAction = actionCreatorFactory<ProcessQueryResultsPayload>(
+  'explore/PROCESS_QUERY_RESULTS'
+).create();
+
+export const runQueriesBatchAction = actionCreatorFactory<RunQueriesBatchPayload>('explore/RUN_QUERIES_BATCH').create();
+
+export const subscriptionDataReceivedAction = actionCreatorFactory<SubscriptionDataReceivedPayload>(
+  'explore/SUBSCRIPTION_DATA_RECEIVED'
+).create();
+
+export const limitMessageRatePayloadAction = actionCreatorFactory<LimitMessageRatePayload>(
+  'explore/LIMIT_MESSAGE_RATE_PAYLOAD'
+).create();
 
 export type HigherOrderAction =
   | ActionOf<SplitCloseActionPayload>
